feat: Added scramble_level arg to pluboard.scramble method
Added a scamble level argument to plugboard.scramble() method. This was done to better emulate the actual enigma machine, as rarley was it the case that all pluboard sockets were commited to a seperate letter, usally only 10 pairs were chosen, and this is echoed in the default value. In order to maintain randomness, the entire list of keys is randomised before looped through

closes #2